Major changes to persistent state passing
Major changes were made to the way persistent state is passed
between mobile_launcher and display_launcher.  The correct way
to specify a file to be used is to send it first with the send_file_1
RPCs and then specify how it is to be used with a file-specific
RPC like load_vm_from_attachment or use_persistent_state.  Then,
only the pathname needs to be identical between send and usage calls
for things to work.